More changes to argument parser
Fixed: If last argument in argument list was one character
long, it would not be picked up.

Breakage: Previously, it was valid to have two tokens side by side
without a whitespace separator if the first token was a quoted string.
This is no longer possible and will throw an exception in the next
release.